Singular Covariance
In NetEst2, we may run into the case where the covariance matrix is
singular. To deal with this we delete a minimum number of
row(i),column(i) pairs.

desing.m does this, using arrayincrement to run through possible things
to delete and delagents to delete specific rows and columns

NetEst2 uses this and handles the case where the covariance matrix is
singular by deleting the minimum number. Then it does the estimate,
fills the deleted agents with zeros and runs the algorithm as usual.

ConnectionTester is running NetEst2 trials for randomized connections.