Singularity Check
This fixes the case where the matrix is singular but has numerical
determine really small or really large.

Singcheck detects what the cond is and if it gives indication of
singularity: cond < 10^-10 or cond > 10^10 or cond = inf, flags the
matrix.

When the matrix is flagged we run the desing.m to remove a minimal
amount of rows.