Ameliorate cong.m Error
Gets rid of the error that was occuring before. For unknown reasons,
matrices with NaN entries were being entered into cong, which doesn't
accept them.

When this happens, this commit makes it so that singcheck is flagged
when this occurs. No errors seems to arise yet, but more investigation
should be done.

So far: no errors for [6 3 1] layer with [2 1] connections.